refactoring - big package with general engine code + math package with symbolic regression stuff
