+++ src/main/scala/Scalatra.scala
@@ -4,8 +4,7 @@ import org.scalatra.LifeCycle
 import org.slf4j.LoggerFactory
 import pl.edu.agh.hubert.configuration.Configuration
 import pl.edu.agh.hubert.datasets.DataSets
-import pl.edu.agh.hubert.engine.EvolutionExecutor
-import pl.edu.agh.hubert.experiments.MemoryExperimentRepository
+import pl.edu.agh.hubert.engine.{MemoryExperimentRepository, EvolutionExecutor}
 import pl.edu.agh.hubert.servlets._
 
 class ScalatraBootstrap extends LifeCycle {
