+++ src/main/scala/pl/edu/agh/hubert/engine/engine.scala
@@ -1,0 +1,8 @@
+package pl.edu.agh.hubert
+
+/**
+ * Created by pkoperek on 9/28/15.
+ */
+package object engine {
+  type Input = Array[Array[Double]]
+}
