+++ src/test/scala/pl/edu/agh/hubert/engine/LanguageTest.scala
@@ -1,10 +1,8 @@
-package pl.edu.agh.hubert.languages
+package pl.edu.agh.hubert.engine
 
 import org.scalatest.FunSuite
-import pl.edu.agh.hubert.hubert.Input
-
+import pl.edu.agh.hubert.engine.LanguageProtocol._
 import spray.json._
-import pl.edu.agh.hubert.languages.LanguageProtocol._
 
 class LanguageTest extends FunSuite {
